Change how the parser parses function
Functions are now parsed at the Number stage, and require
its argument to be enclosed in parenthese. This gets rid
of the awkvard special step for funtions, and makes equations
eaiser for humans to read. It also makes name parsing cleaner
since we now do it all in one place.