ework the way input is read
The input is now read character by character, by setting the
terminal in raw mode. This makes it possible to use the arrow keys
to get previous input, which gives a smoother editing experience.

However there is still some bugs to be worked out - especially
with regard to utf8 - and some edge cases.

It also ONLY works on linux now, which needs to change...