Change error highlighting
When we run in interactive mode, the equation is aleady right above
the error message, so we might as well just highlight this,
instead of printing it again and *then* highlighting.

However this is still needed when the equations are given as
commandline arguments, so an extra argument was added to the
print_location_highlight function which specifies whether we want
to print the input or not.