Rework AST branches
This is in preparation for functions with multiple arguments.

We now store the AST branches in a Vec, with some helper methods
to deal with the most common branch cases. This means that we are
no longer restricted to at most two branches, which will help make
the code more general.

Some cleanup of the old code to take advantage of this new feature
might be needed.